conf: drop TODOs re error reporting and etcd rev
We'll work on the error reporting during re-reading+re-applying
implementation. This task is tracked and there is no need to highlight
it in random places across the code. Moreover, I guess that pcall'ing at
proper places should be enough: we anyway can't replace error reporting
function inside, say, box.cfg().

Regarding etcd revisions: same, the task is tracked and we'll work on
it.