conf: add schema.array()
conf: drop implemented idea comments from schema

conf: drop <schema>:equals()

It is not used for now.

conf: make <schema>:merge() schema-aware

The main reason is to distinguish a record from an array. We should
perform a deep merge for a record, but not for an array.

In fact, the previous implementation was buggy not even counting the
point regarding arrays.

```
 1 merge_impl = function(a, b)
 2    if type(a) ~= 'table' then
 3        return b
 4    end
 5    if b == nil then
 6        return a
 7    end
 8    assert(type(b) == 'table')
 9    local res = {}
10    for k, v in pairs(a) do
11        res[k] = merge_impl(v, b[k])
12    end
13    <...>
```

If `v` is non-nil and `b[k]` is `nil` on line 11, then the
`merge_impl()` call will return `nil` anyway due to the if-condition on
line 2.

conf: drop unused 'required' annotation

conf: support allowed_values annotation

conf: add schema.enum()

conf: use schema.enum() instead of schema.scalar()

conf: add schema.set()

conf: add validation for union of records

conf: copy the path in :filter() and :walkthrough()

The `ctx.path` is a mutable table -- it changes during the traversal.

It is usually ok, because :walkthrough() just calls a usr provided
callback. There was an assumption that all the data will be processed in
the callback.

However, it is not the case for :filter(). It accumulates returned data
and returns an iterator on them.