ci: add required deps for testing with Clang
Add required dependencies to make the `egor-tensin/setup-clang` GitHub
action work. This action is planned to be used in the next commits for
setting Clang up on the runners and requires `powershell` be installed.
So adding the `snap`, `snapd`, `powershell` packages. The first two are
needed for `powershell` installation.

Part of tarantool/tarantool-qa#317
Part of tarantool/tarantool-qa#318
Part of tarantool/tarantool-qa#319

NO_DOC=ci
NO_TEST=ci
NO_CHANGELOG=ci